Converts button to typescript (#80194)
Converts one of our most used components into TypeScript. Fixes an issue
with it where ellipses were not working.

This is tricky though. Some people are using ellipsis prop but passing
in full nodes. Tgui is not smart enough to handle this - it only works
on strings! You will need to custom write the style if you're trying
this. I've wrote up documentation in the component and a restrictive
type that will warn anyone trying to concatenate a node.
Closes #78220
N/A nothing player facing