Input changes take XII: Responsive small lists (#80720)
## About The Pull Request
Admittedly it's been tough to find a sweet spot that will let us handle
large lists and basic inputs. This PR aims to address this by adding the
`expensive` prop on Input. This introduces the debounce that #80688
without compromising the responsiveness of smaller filters like those in
orbit, giving the author the keys to monitor performance of their UIs.
On a more human note inputs just feel better with this PR.

Along with this change I've expanded documentation for the typescript
types and included a discriminating union to let contributors know that
the `expensive` prop requires `onInput` to work.
## Why It's Good For The Game
Bug fixes, responsiveness, etc
## Changelog
:cl:
fix: Search bars for smaller lists should return to their former
responsiveness.
/:cl: