tgui: Strings.ts (#83096)
Wrote better docs and tests for string helpers in tgui.

Something I noticed, which I'm ok with reverting, is that multiline does
precisely nothing. And despite being a template string fn, it doesn't
accept vars for interpolation. It looks as if it were used to circumvent
a babel limitation, which we are no longer using.

Removed:
- babel plugin (unused)
- multiline (no impact)
- createGlobPattern (unused)
- buildQueryString (unused)

<details>
<summary>Spot the difference</summary>

![Screenshot 2024-05-06
095420](https://github.com/tgstation/tgstation/assets/42397676/bbf2f58e-81b7-4023-8da4-8d7f19cd3bb1)

![Screenshot 2024-05-06
095443](https://github.com/tgstation/tgstation/assets/42397676/24f303ba-f352-4132-a2a6-873dff0d979b)

(There is none)

</details>

More core tools in TS.
Better docs so hopefully these functions are more widely used (and done
so properly)
Tests so things don't break