Add a choice to keep being a vampire after a embrace or after being turned into a ghoul (#398)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

After being transformed into a vampire/ghoul, now you can pick between
"Yes" or "No" to save that change on your slot, keep in mind that WL
clans for those without the WL, will be changed to their non WL versions
or randomized.
For power gaming reason, if your sire is below the 13 gen, you will
always be a 13th gen as well, if they are 13th gen you will be turned
into a Caitiff 14th gen but unfortunately the game by default will force
the 14th gen to turn 13th on the char slot.

On a side note for coders, create_disciplines does not work if you pass
a list, i tried a couple of time to fix but it was not my focus so i
used give_discipline directly, that also means that gargoyles created
ingame don't recieve their disciplines.

Vamp choice:

![vamp_ask](https://github.com/user-attachments/assets/f2916695-312e-4b62-a44f-5746a3eecfdb)

Ghoul choice:

![ghoul_choice](https://github.com/user-attachments/assets/9aacb702-c615-4d3f-ae25-d14ac8591923)

## Why It's Good For The Game

This is mainly for rp reasons, so people call roleplay as real
sire/child or turn a human into a ghoul and they choose to keep,
also WLs can make their story based of a real player as their sire and
that can happen mechanically.
Also someone who plays human alot and spent alot of xp on it can turn
into a vampire/ghoul without the need to reset their stats.

## Changelog
:cl:
Add: A choice to keep the vampire change on your slot after being
transformed
Add: A choice to keep the ghoul change on your slot after being
transformed.
Fix: The child of a embrace now gain their skills
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->