Banu Haqim Haven (READY FOR MERGE/REVIEW ! ! ! !  ) (#463)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

This PR adds a quirky little coffee house that serves as a haven for the
Banu Haqim clan. I have done my Best what with our Limited Assets and
All That.

Below will be some pictures.

![image](https://github.com/user-attachments/assets/ee86b9d8-f26b-4fb4-8392-d3163e3dc579)

![image](https://github.com/user-attachments/assets/dca28789-0485-4cbf-ab3a-0658783acb68)

![image](https://github.com/user-attachments/assets/80e8f4d6-5875-47d6-af6d-13f60b4b6f2c)
(Yes I fixed the missing space tile in this last image lol)

![image](https://github.com/user-attachments/assets/1dcbce7b-d1be-4aa6-82d3-f98e40dea9fd)

![image](https://github.com/user-attachments/assets/f783be52-981b-4dc5-90f5-489932a96a3f)

![image](https://github.com/user-attachments/assets/7af69c39-47ee-46f8-8cf2-093725fcda84)

![image](https://github.com/user-attachments/assets/3ff284ae-48ed-4043-9403-72b842bbac2f)
I've added some more lights in following commits but it's basically the
same. And yes I added keys for this of course.

## Why It's Good For The Game

Banu Haqim are (or should be) the most numerous whitelisted clan imo, so
it only makes sense for them to get a Haven. Also, I like Banu Haiqm, so
I am Incredibly Biased. Fuck You. Also Coffee Houses Are Cool Or
Something.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl:
add: Added Banu Haqim Haven, in the form of a Coffee House.
add: Added keys for said Haven.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->