Readjusts values & lessens the amount of corpsed bodies (#516)
## About The Pull Request

The following pr makes it so the total cost of one NPC is roughly 2K if
you trade in every single organ (Going by the cost = value). This PR
also significantly buffs drug selling and weed trading as well!
Encouraging players to set up farms and defend them as a means of
establishing territory.

## Why It's Good For The Game

As mentioned above.

## Changelog
:cl:
Lowers the value of all organs to equal 2K.
Increases the value of weed to 350, meth to 650, and cocaine to 750.
/:cl: