Clinic update (#511)
## About The Pull Request

This is just a tiny-tiny remap of the clinic trying to make it more
pleasurable to actually be in and play.
Made Clinic Director so there's at least some authority in the clinic to
keep things clean and so on.

![image](https://github.com/user-attachments/assets/8b8a1c2d-2341-4c70-b88a-f55992999a84)

![image](https://github.com/user-attachments/assets/d220960d-efa2-4e81-be42-ef99a897972c)

Clinic Director's labcoat! BIG THANKS TO FLAVRIUS <3

![image](https://github.com/user-attachments/assets/f65d16df-fef5-4e7f-85ba-d70512e3052a)

## Changelog

Added some stuff that can't be seens from the pictures:
— Clinic Director and his new labcoat!
— epipen box, armbands, additional EC radio, two freezers with organs
(and copypasted them from tg), a small canteen with some food, psiocidin
pills in the basement and in psychiatrist's office (to larp bout
antdeprisants and so on)
— removed xenoblood from the bloodfreezers
— removed og radio from the doctor's outfit (we have R25 thingy)

---------

Co-authored-by: Joshua Kidder <49173900+Metekillot@users.noreply.github.com>
Co-authored-by: Metekillot <49173900+Metekillot@users.noreply.github.co>