Adjusts drug money for a final time (#568)
## About The Pull Request

The original PR I made for this was a bit too overtuned; I've since
rectified this with this PR here. Fewer people are of the dreaded "path
of the organ harvester" so that worked at the very least! Anyway, heres
the math:

350 x 4 (meth) for 1400,
An epi bottle is 200 dollars (vendor price)
Potassium iodide is 100 (vendor price)
Gasoline canister is 250 (vendor price), but I misread the code, its now
taking up 100 out of 1000 so you are really only going to get 10 batches
of meth.

Adding it together, each batch at the cost they are sold for will yield
a profit of 850 per batch and making 10 of it is 8500 but this requires
significant cost investment, unlike weed which requires roughly 15
minutes to grow from round start.

150 for weed, previously 50.
You'll generally hit 900 dollars if you get six leaves as a best case
scenario from RNG (I might be wrong here), the benefit being that you
can maintain this so long as you water the plants.

## Why It's Good For The Game

People had too much money and it was kind of crazy and I overdid some
values, this accounts for 5 social however so do be aware.

## Changelog
:cl:
Adjusts meth price.
Adjust weed price
/:cl: