Four turtlenecks, Tremere clothes, and a new robe. (#664)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds in four variations of a turtleneck (you can buy these in
stores)
A Tremere robe, a Magister robe, and a "red trenchcoat" (a strauss
trenchcoat), none of which are obtainable in this PR and need to be
mapped in by @AllCorruptingWyrm in #656 . Merge my PR first so they can
add these items to the map.
Also adds a new blue robe, which is just a recolor of the other ones.
(You can buy it in the shops)

All of these sprites were made by Eluxor on the discord and were given
to me to be added, barring the blue robes, which were made by
Atpengineer and given to me to be added.

![image](https://github.com/user-attachments/assets/b4ccde28-55ab-4205-a50e-7c91ddc7d8fe)
The items in-game, fully tested.

![image](https://github.com/user-attachments/assets/f8112c38-c77b-4326-b514-e9e43d305204)
Pictures of some of the sprites in their raw asset file

![image](https://github.com/user-attachments/assets/822a34e8-6e3b-475a-a2bf-f724299d745e)
Some of the new sprites lined up against our existing ones, to show how
they blend in

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More clothing variety is good.
Tremere are honestly kind of boring and need more of their own unique
look. So, now, if they want to wizardmaxx, they can do so.
Blue robes are cool too.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Four turtlenecks
add: Two tremere robes and a regent trenchcoat
image: turtleneck_white, turtleneck_black, turtleneck_red,
turtleneck_navy, tremere_robes, magister_robes, strauss_coat,
robes_blue, etc etc
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->