Small Music Update (#435)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This PR adds a whopping **48** songs to the jukebox from the 90s, and
the top 100 of '02,'03,'04, and a few songs from '05.
## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
Jukeboxes have too limited of a selection, and it's a shame. With this,
areas can feel somewhat more than simply "yeah we're in the bar" with
more music. That, and people keep complaining about Punk Tactics.

Now go blare Milkshake at 100 volume, you cretin.
## Changelog
:cl:
add: Added 62 period-appropriate songs to the jukebox of varying flavor
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->