The Flaggening (Also; language icons) (#670)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds flags to join the American flag (not pictured) we already have.

![image](https://github.com/user-attachments/assets/c2fca3bb-5463-4eb6-9564-cf9865655425)

Also, standardizes language icon colours and sizes, and replaces the
current icon of Japanese (temple gate) with the flag of Japan.

## Why It's Good For The Game

The new flags easily fit into the game, and it's nice to have options.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Flag of California, of Great Britain, of Germany, of Spain, of
Italy, of the Vatican, of Russia, of the USSR, of China, of Taiwan, and
of Japan
add: Language icon (Japanese flag)
edit: Standardizes language icon colours and sizes
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Adri <33333517+Miraviel@users.noreply.github.com>