Nerfs diablerie (#442)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

You no longer get reset to generation 13 by diablerie, as this is an
extreme tool to allow (hypothetically) just about anyone to use on
anyone they can robust.
In addition, diablerie now simply lowers your generation by 1 on a
successful attempt **(edit: when performed on someone of a lower
generation),** or by 2 if the victim was at least 4 generations below
you. _Failed_ diablerie raises your generation by 1.

## Why It's Good For The Game

The server is getting rid of many of the progress-loss features the
codebase once had, and I think this is appropriate for that direction.
The possibility of losing 10+ hours of XP after losing a fight might
have suited what the original server was going for, but not this one,
and it's not practical to effectively moderate every single use.
This may even make diablerie more appealing, now that you're not causing
a significant blow to an actual players's progress. Being evil is fun
when characters suffer, not when _people_ lose a major chunk of XP
through potentially no fault of their own.

## Changelog
:cl:
balance: Diablerie no longer affects the victim's generation. Failed
diablerie raises the attacker's generation by 1.
balance: Diablerie will only lower your generation by 1, or by 2 if the
victim is 4 or more generations below you.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->