Fixes rat-caused runtime which means you can store and throw them from your hand again & Animalism 5 fix. (#619)
## About The Pull Request
When a rat is not "in view" of anything it deletes itself to not cause
mass-rat-event.
If you pick a rat up, it is not "in view" and deletes itself, but the
mob-holder code will still have it in your hand or in your bag until you
decide to drop it in which case it runtimes, because forcemove is given
a null destination, and then deletes the rat.
This adds a check so that, if the rat is technically not anywhere, the
rat won't be deleted inside things like crates or backpacks.
Also changes oviewers to viewers so that it won't try to delete itself
if its viewed by the person holding it.
Added a check for if the rat is controlled by a client (like with
current animalism 5) in which case we don't want to delete it suddenly.

Credits for the help needed to identify, diagnose, and patch this issue
with the people in the discord: silencer_pl, adri, and steve5315
## Why It's Good For The Game
Being able to store a mass quantity of rats in your backpack and being
able to throw them at people is a hallowed tradition for nosferatu and
it is truly a shame that the current state of things disallows such
practice. Also runtimes are icky.

## Changelog
:cl: Vondiech/Citruses
fix: You can now pick up, store in bag, and throw rats again without
them disappearing.
fix: Animalism 5 rat-shapeshifting no longer gets immediately undone
when out of view.
/:cl: