Removes Extra Lattices Causing Runtimes. (#710)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes a double lattice in Pentex that is causing runtimes on Z1.

![image](https://github.com/user-attachments/assets/ea6ea58f-a8be-41ce-a1f2-49e4c191f384)

## Changelog

:cl:
map: Removed double lattices in Pentex causing runtimes.
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->