Adds Pentex Clothing (#720)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Sprites provided by [Eluxor](https://ko-fi.com/eluxor) and commissioned
by Menkhasa.

Adds a variety of Pentex clothing (named after subsidiaries for IC
purposes), including:

- Labcoat
- Suit / Suitskirt
- Executive Suit / Suitskirt
- Shortsleeve Shirt
- Longsleeve Shirt
- Turtleneck
- Janitor Jumpsuit
- Yellow/White Hardhats
- Balaclava
- First Squad Beret

https://github.com/user-attachments/assets/f5e022cc-0ea2-4ee8-9472-cab696927f1a

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Additional clothing options are always good for variety in the game, and
with the addition of a large Pentex facility that will eventually have
jobs within, the clothes will be useful to set them apart/make them
distinctive.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: A variety of Pentex-themed clothing.
code: Created /pentex types for the various clothing categories, as
opposed the /vampire type.
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Adri <33333517+Miraviel@users.noreply.github.com>