Prevent stuns on vampire bite (#688)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

This PR turns the sleep check at the end of the bite code into a proper
signal to determine if whomever was bitten has resisted the bite or not
once the bite is over. If they have, they are not put to sleep, if they
haven't, they are put to sleep for 5 seconds.

This also makes it so _only_ NPCs have a guaranteed constant stun for
each cycle of the bite, giving garou and other players, even humans, the
chance to escape. The difference is humans will pass out after breaking
free, supernaturals who are meant to be resistant will not. This still
leaves humans at the extreme disadvantage they should be under when
going against vampires, but at least lets them be a little nuisance if
unwilling and strong willed.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

There's been plenty of cases where werewolves, even in Crinos, have been
chain-stunned by bite spamming vampires. This is both non lore compliant
and also unfun in general to suffer. This will remove the issue
completely, add a new signal to process bite resistances with and give a
little buff to kine as a treat because human mains suffer enough as is.
It also adds kuei jin to the list of things that should be 'resistant'
to the kiss.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: makes garou immune to vampire bite stun
balance: makes kuei-jin immune to vampire bite stun
balance: allows humans to resist out of a bite grab, will still KO them
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Joshua Kidder <49173900+Metekillot@users.noreply.github.com>