Lavaland/Penumbra ashstorms removal (#782)
## About The Pull Request
Deletes most if not all codes related to lavaland ashstorms as well as a
few other bits of lavaland code thats not really relevant to how
lavaland is currently used.
Deletes a magic loot staff that was used only to summon ashstorms, since
that is completely pointless and much more of a hassle to try and change
now that ashstorms are gone.

Honestly most of the lavaland code will need to be gutted sooner or
later but for now this will have to do as a small but somewhat front-end
part of it.
## Why It's Good For The Game
Ashstorms are another system thats only superficially used at the moment
to make noise in the penumbra/lavaland, right now we have absolutely no
use of this besides making the penumbra noisy in a way irrelevant to how
we're using it.

## Testing Photographs and Procedure
Proof of compilation:

![2025-02-17_10-59-33](https://github.com/user-attachments/assets/804bda84-bc6e-4b14-b013-2d4f9d567e30)
Proof of no runtiming

![dreamseeker_2025-02-17_11-06-08](https://github.com/user-attachments/assets/819fc0af-9fc8-406f-9313-e9872e06f26f)

</details>

## Changelog

:cl: Vondiech/Citruses
del: Removed the lavaland/penumbra ashstorm weather datum that was
causing the ashstorm sounds to play.
/:cl: