Hand sickle (#730)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds a 30 Force melee weapon with slightly more armor pen than a knife,
no block chance, and a custom attack_verb (reaps). It is a 2x1 grid
item, and fits on a belt.
Also places it in a few locations; The Wendigo encampment, the Graveyard
Keeper's armory, under the Watchtower, and the closets that do not have
a spare scythe.

![image](https://github.com/user-attachments/assets/d62a04b0-b075-405c-be43-b6217d75143b)

![image](https://github.com/user-attachments/assets/1a715994-3572-41de-9e0c-18b44ed154dc)

![dreamseeker_57zRjhNznA](https://github.com/user-attachments/assets/c229ca5d-942f-4e35-a203-1e8c07730efa)

![dreamseeker_qrPisAFqjH](https://github.com/user-attachments/assets/3409bbc1-079f-4776-8892-dc99156608f1)

![dreamseeker_PFW4oiud6x](https://github.com/user-attachments/assets/36a645a4-0308-4cda-9f3b-b98c4ec954f8)

![dreamseeker_PeUIowmAa0](https://github.com/user-attachments/assets/ad85bee1-8054-4ff1-99db-7f01a1881c88)

![dreamseeker_YXg2l1m0XG](https://github.com/user-attachments/assets/e4da7296-0045-48ed-a3de-ee59bd5b8296)

![dreamseeker_0r3XxZEctb](https://github.com/user-attachments/assets/32f8c384-3b33-4807-b096-1f5aae688bb2)

![dreamseeker_DxdZZPxiBV](https://github.com/user-attachments/assets/8dd33970-dd36-471e-aeb5-61ac9fee20bf)

![image](https://github.com/user-attachments/assets/a311f463-63aa-4dd7-8e27-056467603388)

![StrongDMM_wTUNimRKrT](https://github.com/user-attachments/assets/cd921842-0217-4be4-a86f-39cd1dd61c40)

![StrongDMM_h4MCAuZh4a](https://github.com/user-attachments/assets/aeb2f08d-4646-4761-ba4d-d367bddd8fcd)

![StrongDMM_tHMpd61AeK](https://github.com/user-attachments/assets/d00dc4d4-772a-42c6-876e-210282765f26)

![StrongDMM_XKORrdWibI](https://github.com/user-attachments/assets/bc7f2895-56ec-45ae-9ae8-a38a036e142e)

## Why It's Good For The Game

More thematic weapons with minor differences let players diversify their
options. A hand sickle is a trope heavy horror weapon that can help many
scenes, in a small way.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Hand Sickle
add: Map Spawns for the Hand Sickle
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->