Allows Malkavians to speak in VTM:B style fonts (#754)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Makes Malkavians text resemble the Bloodlines Malkavian speech font.
This PR utilizes the existing comic sans font as well as some specific
unicode replacement characters on certain letters to achieve something
resembling the RPG's font.

It is toggleable via an action button, ala hivemind ability.

This PR is an atomization from #722, she did all the work, I just
refactored to work with actions.dm instead of special organs
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Malkavians are currently hard to distinguish from non-Malkavians because
they suffer no real penalty to choosing to be the insane clan aside from
the occasional harmless hallucination. In Bloodlines, you're a raving
lunatic with a variety of hallucinations that directly impact your
ability to both be understood and to understand the world around you.
This PR brings us closer to that, where Malkavians' dialogue now seems
as unhinged as their characters should come across.

The font styling starts disabled, and can be turned on in a new
'Malkavian' verb tab after the Malkavian spawns. Good for if your
Malkavian is good at masking it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: buffyuwu, singul0
add: Malkavians can now use VTM:B styled fonts
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: buffyuwu <buffyvrc@gmail.com>