Properly removes Diablerist quirk (#625)
## About The Pull Request
Properly removes the Diablerist quirk from the game, which showed up as
"Test Quirk" in the quirk menu.
It did not have any permanent consequences, it merely marked the
character as a diablerist for the duration of a round.

## Why It's Good For The Game
It's a bugfix. It was intended to be deleted months ago.

## Changelog
:cl:
fix: Properly removed the Diablerist quirk from the game.
/:cl:

---------

Co-authored-by: Joshua Kidder <49173900+Metekillot@users.noreply.github.com>