Makes the cost of raising your Dharma level scale with level (also, a non-critical but substantial bug fix) (#893)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Previously, Kuei-jin had to spend 20 XP for each Dharma level. This PR
changes that cost to 5 XP per current level to a maximum of 20, so the
costs are 5, 10, 15, 20, 20.
Also fixes the way Kuei-jin max HP is calculated, because this didn't
properly take both physique and dharma into account.

## Why It's Good For The Game

The benefits of Dharma levels 2-4 are limited to +25 HP, and 20XP is
just a crazy cost for that at the moment - Physique does that and more!
Only 5 and 6 provide any other benefit, those being other ways to draw
chi (very useful) and an increase to your maximum resources.

## Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<!-- You can uncomment line 1 @ _maps/_basemap.dm to boot up a test map
that loads much faster. -->
<details>
<summary>Screenshots&Videos</summary>

![image](https://github.com/user-attachments/assets/cbbfa82e-5748-4731-b486-c43c3620ad5e)
I have 935 XP.

![image](https://github.com/user-attachments/assets/f812f3b5-55e0-4830-8be5-4997c054d559)
I now have 925 XP.

![image](https://github.com/user-attachments/assets/765f743f-49c8-43d3-97bc-119db69ef858)
The above image was taken with 6 physique and Dharma 3.

![image](https://github.com/user-attachments/assets/40ae3033-6e83-44a0-85dc-d6f0d0ad3b54)
This one was taken with 5 physique and Dharma 6.
</details>

## Changelog

:cl:
balance: kuei-jin only need to pay 5, 10, and 15 XP for the first three
dharma rises
fix: kuei-jin should have the correct HP for their level and stats
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->