Prevents protected audio from crashing the server with no survivors (#90975)
Fixes #90694

I believe the heart of this issue is that the audio player was trying to
log an error object, but logs only take strings. This should prevent the
blue screen and provide admin feedback
- Everyone survives your fatal attempt to play king gizzard the lizard
wizard
- Better admin feedback why you cannot play aforementioned song
:cl:
fix: We're no strangers to bugs: Playing protected audio shouldn't crash
the server anymore!
/:cl: